Update passing data between calendar controller and view + renamed WebClinic to WebClinicGUI in namespaces
