+++ WebClinicGUI/Controllers/AddAppointmentController.cs
@@ -1,15 +1,12 @@
 ﻿using System;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
-using WebClinic.Data;
-using WebClinic.Models.Users;
-using Microsoft.EntityFrameworkCore;
-using WebClinic.Models;
 using System.Collections.Generic;
+using WebClinicGUI.Models.Users;
+using WebClinicGUI.Models;
 
-namespace WebClinic.Controllers
+namespace WebClinicGUI.Controllers
 {
     public class AddAppointmentController : Controller
     {
