+++ WebClinicGUI/Controllers/PhysiciansController.cs
@@ -1,11 +1,7 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using WebClinic.Models;
+﻿using Microsoft.AspNetCore.Mvc;
+using WebClinicGUI.Models;
 
-
-namespace WebClinic.Controllers
+namespace WebClinicGUI.Controllers
 {
     [Route("[controller]/[action]")]
     //[Authorize(Roles = "Receptionist,Physician")]
