+++ WebClinicGUI/Models/PatientViewModel.cs
@@ -1,11 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
-using WebClinic.Models.Users;
+using WebClinicGUI.Models.Users;
 
-namespace WebClinic.Models
+namespace WebClinicGUI.Models
 {
     public class PatientViewModel
     {
