+++ WebClinicGUI/Models/Users/AppUser.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace WebClinic.Models.Users
+namespace WebClinicGUI.Models.Users
 {
     public enum Sex
     {
