+++ WebClinicGUI/Models/Users/Receptionist.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebClinic.Models.Users
+namespace WebClinicGUI.Models.Users
 {
     public class Receptionist : AppUser
     {
