+++ WebClinicGUI/Models/Users/Role.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace WebClinic.Models.Users
+﻿namespace WebClinicGUI.Models.Users
 {
     public class Role
     {
