+++ WebClinicGUI/Program.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Hosting;
 
-namespace WebClinic
+namespace WebClinicGUI
 {
     public class Program
     {
