+++ WebClinicGUI/Services/INetworkClient.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
 
@@ -8,7 +6,7 @@ namespace WebClinicGUI.Services
 {
     public interface INetworkClient
     {
-        public static Uri ServerUrl = new Uri("http://localhost:51210/api/");
+        public static Uri ServerUrl = new Uri("https://localhost:44359/api/");
         public static string ApiName = "API";
         public Task<T> SendRequestAsync<T>(HttpMethod httpMethod, string endpoint, StringContent content = null);
         public Task<T> SendRequestWithBodyAsync<T>(HttpMethod httpMethod, string endpoint, T body);
