+++ WebClinicGUI/Startup.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Globalization;
 using Microsoft.AspNetCore.Authentication.Cookies;
@@ -11,10 +10,9 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
-using WebClinic.Data;
 using WebClinicGUI.Services;
 
-namespace WebClinic
+namespace WebClinicGUI
 {
     public class Startup
     {
