+++ WebClinicGUI/Views/Account/Register.cshtml
@@ -1,9 +1,9 @@
-﻿@using WebClinic.Models.Users
+﻿@using WebClinicGUI.Models.Users
 @using Microsoft.AspNetCore.Mvc.Localization
 
 @inject IViewLocalizer Localizer
 
-@model WebClinic.Models.RegisterViewModel
+@model WebClinicGUI.Models.RegisterViewModel
 @{
     ViewBag.Title = @Localizer["Registration"];
 }
