+++ WebClinicGUI/Views/AddAppointment/Index.cshtml
@@ -1,9 +1,9 @@
-﻿@using WebClinic.Models.Users
+﻿@using WebClinicGUI.Models.Users
 @using Microsoft.AspNetCore.Mvc.Localization
 
 @inject IViewLocalizer Localizer
 
-@model WebClinic.Models.AddAppointmentViewModel
+@model WebClinicGUI.Models.AddAppointmentViewModel
 @{
     ViewBag.Title = @Localizer["New appointment"];
 }
