+++ WebClinicGUI/Views/Patients/Edit.cshtml
@@ -1,9 +1,9 @@
-﻿@using WebClinic.Models.Users
+﻿@using WebClinicGUI.Models.Users
 @using Microsoft.AspNetCore.Mvc.Localization
 
 @inject IViewLocalizer Localizer
 
-@model WebClinic.Models.PatientViewModel
+@model WebClinicGUI.Models.PatientViewModel
 @{
     ViewBag.Title = @Localizer["Edit patient"];
 }
