+++ WebClinicGUI/Views/Physicians/Edit.cshtml
@@ -1,9 +1,9 @@
-﻿@using WebClinic.Models.Users
+﻿@using WebClinicGUI.Models.Users
 @using Microsoft.AspNetCore.Mvc.Localization
 
 @inject IViewLocalizer Localizer
 
-@model WebClinic.Models.PhysicianViewModel
+@model WebClinicGUI.Models.PhysicianViewModel
 @{
     ViewBag.Title = @Localizer["Edit physician"];
 }
