+++ WebClinicGUI/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
-﻿@using WebClinic
-@using WebClinic.Models
+﻿@using WebClinicGUI
+@using WebClinicGUI.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
